src/requirements.txt
coreapi==2.0.8
Django==1.10.1
django-crispy-forms==1.6.0
django-filter==0.15.2
djangorestframework==3.4.7
gunicorn==19.6.0
itypes==1.1.0
Markdown==2.6.7
mysqlclient==1.3.8
Pillow==3.3.1
requests==2.11.1
uritemplate==3.0.0
